Minor fix and added Allow-Control-Allow-Header in CORSFilter

+++ src/main/java/com/kunleawotunbo/gameplay/dao/WeeklyGamesAnswersDao.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 public interface WeeklyGamesAnswersDao {
     
-    WeeklyGamesAnswers findById(int id);
+    WeeklyGamesAnswers findById(Long id);
 
     boolean saveWeeklyGamesAnswer(WeeklyGamesAnswers game);
